requirements.txt
Django==3.2
django-filter==21.1
djangorestframework==3.12.4

PyMySQL==1.0.2

